Refactor:HomeTopBar 아토믹 디자인 적용
