+++ app/common/assets/icons/search/SearchIcon.tsx
@@ -1,5 +1,6 @@
 import { IconProps } from "../icon-type";
 
+//Atom
 export default function SearchIcon({ className, onClick }: IconProps) {
   return (
     <svg
