+++ app/common/components/ui/topBar/HomeTopBar/HomeTopBar.tsx
@@ -1,4 +1,4 @@
-import HeaderText from "@common/assets/icons/header/HeaderText";
+import HeaderTextIcon from "@common/assets/icons/header/HeaderTextIcon";
 import SearchIcon from "@common/assets/icons/search/SearchIcon";
 import Button from "@common/components/ui/buttons/Button/Button";
 import Link from "next/link";
@@ -7,11 +7,12 @@ interface HomeHeaderProps {
   showBtn: boolean;
 }
 
-export default function HomeHeader({ showBtn }: HomeHeaderProps) {
+//Organism
+export default function HomeTopBar({ showBtn }: HomeHeaderProps) {
   return (
     <header className="fixed flex justify-between items-center px-[2rem] pt-[1.6rem] pb-[1rem] w-full z-10 bg-[#F5F8FF]">
       <div className="py-[0.7rem]">
-        <HeaderText />
+        <HeaderTextIcon />
       </div>
       <div className="flex items-center">
         {showBtn && (
@@ -20,7 +21,6 @@ export default function HomeHeader({ showBtn }: HomeHeaderProps) {
               pathname: "/search",
               query: { keyword_search: true },
             }}
-            scroll={false}
           >
             <Button className="w-[12.5rem] h-full py-[0.6rem] mr-[1.2rem] body2-semibold">
               키워드로 공간 찾기
